ReReSolved

+++ app/main.py
@@ -8,11 +8,15 @@ def move_file(command: str) -> None:
     if len(split_command) != 3:
         return
     cmd, source_path, dest_path = split_command
+    if not os.path.exists(source_path) and os.path.isfile(source_path):
+        return
     if cmd != "mv":
         return
+    final_dest_path = os.path.join(dest_path.rstrip(os.path.sep))
+    source_path = os.path.basename(source_path)
     if not os.path.dirname(dest_path):
-        with (open(source_path, "r") as file_src,
-              open(dest_path, "w") as file_dest):
+        with (open(source_path, "rb") as file_src,
+              open(final_dest_path, "wb") as file_dest):
             file_dest.write(file_src.read())
         os.remove(source_path)
         return
@@ -24,7 +28,7 @@ def move_file(command: str) -> None:
             current_path = os.path.join(current_path, part)
             if not os.path.exists(current_path):
                 os.mkdir(current_path)
-    with (open(source_path, "r") as file_src,
-          open(dest_path, "w") as file_dest):
+    with (open(source_path, "rb") as file_src,
+          open(final_dest_path, "wb") as file_dest):
         file_dest.write(file_src.read())
     os.remove(source_path)
